Add manpages & README documentation
Credit to the swupd developers is due for demonstrating how to easily
integrate the documentation into the build system. While this was
previously hooked up in the makefile target, this change and pending
release now adds the completed documentation to complete the install
target, making abireport now suitable for packaging.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>