libabi: Ignore small files that couldn't be a proper ELF file
Before we examine an ELF file, we first check for the presence of a
valid ELF header, which is 4 bytes. To avoid unnecessarily checking files
that are empty/tiny, and spamming the console with unnecessary text, we
mandate that an interesting ELF file must be at minimum, 5 bytes in size.

While the reality is that an ELF is always going to be bigger than this,
it enforces a sane constant that doesn't break ELF detection, but does
avoid checking useless files that couldn't header match and contain
interesting content at the same time.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>