explode/eopkg: Use a portable function to explode the eopkg
This function is much like the RPM explosion helper, by using a simple
pipe to get directly to the contents of the package. Here we simply
pipe unzip to tar xf, explicitly passing the `install.tar.xz` payload
to tar. Internally an eopkg is simply a ZIP file with a tarball payload
and some XML data, so this is always going to work.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>