walker.go: fix a race condition
There were occasional hangs detected when processing
empty packages. These were the result of a race condition:

a.JobProcessor and the go function with filepath.Walk were executed
concurrently. For empty packages the go function could terminate
before a.JobProcessor loop dispatched all a.nJobs.

When filepath.Walk terminated, it closed a.JobChan.
If the channel was closed during the jobProcessor dispatch loop,
less than a.nJobs were dispatched. However, jobDone expects
to be called exactly a.nJobs times before closing a.storeChan.

The remedy is to ensure that we close a.JobChan after
jobProcessor finished.

Signed-off-by: Juro Bystricky <jurobystricky@hotmail.com>